+++ src/gtt_application_window.c
@@ -33,12 +33,12 @@
 #include "app.h"
 #include "gtt_current_project.h"
 #include "gtt.h"
-#include "log.h"
+#include "gtt_log.h"
 #include "menucmd.h"
 #include "gtt_menus.h"
 #include "notes-area.h"
 #include "gtt_preferences.h"
-#include "props-proj.h"
+#include "gtt_props_dlg_project.h"
 #include "gtt_timer.h"
 #include "gtt_toolbar.h"
 #include "gtt_util.h"
