+++ src/gtt_log.h
