+++ src/gtt_props_dlg_interval.h
