+++ src/gtt_props_dlg_project.c
@@ -19,13 +19,14 @@
 
 #include "config.h"
 
+#include "gtt_props_dlg_project.h"
+
 #include <glade/glade.h>
 #include <gnome.h>
 #include <string.h>
 
 #include "gtt_help_popup.h"
 #include "gtt_project.h"
-#include "props-proj.h"
 #include "gtt_util.h"
 
 typedef struct _PropDlg
