+++ src/gtt_props_dlg_task.c
@@ -18,13 +18,14 @@
 
 #include "config.h"
 
+#include "gtt_props_dlg_task.h"
+
 #include <glade/glade.h>
 #include <gnome.h>
 #include <string.h>
 
 #include "gtt_help_popup.h"
 #include "gtt_project.h"
-#include "props-task.h"
 #include "gtt_util.h"
 
 typedef struct PropTaskDlg_s
